+++ packages/backbone_cli/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.1-dev.2
+
+- Server now properly uses the logging middleware
+
 ## 0.0.1-dev.1
 
 - Initial release as a split up package.